requirements.txt
streamlit >= 1.20.0
torch >= 1.8.2+cu111
pandas >= 1.5.3
numpy >= 1.22.4
pillow >= 9.2.0
ultralytics >= 8.0.118
opencv-python >= 4.6.0.66